Start of Nvidia Physx

+++ SlateEngine/Engine/Core/EngineConfig.h
@@ -16,4 +16,4 @@ constexpr const char* META_FILE_EXT = ".smeta";
 constexpr int PHYSX_VD_DEFAULT_PORT = 5425;
 constexpr const char* PHYSX_VD_DEFAULT_HOST = "127.0.0.1";
 
-constexpr const int GRAVITY_SCALE = -9.81f;
+constexpr const float GRAVITY_SCALE = -9.81f;
